+++ firmware/svc-adc.h
@@ -1,6 +1,6 @@
 #ifndef __SVC_ADC_H__
 #define __SVC_ADC_H__
 
-extern volatile uint16_t adc[];
+extern volatile int adc[];
 
 #endif
